Rename app, reconfigure settings & database
