+++ apps/recipes/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class RecipeApiConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'apps.recipe_api'
+    name = 'apps.recipes'
