+++ apps/recipes/models/__init__.py
@@ -1,0 +1,4 @@
+from apps.recipes.models import category
+from apps.recipes.models import recipe
+from apps.recipes.models import ingredient
+from apps.recipes.models import instruction
