+++ apps/recipes/serializers.py
@@ -1,7 +1,7 @@
 import bleach
 from rest_framework import serializers
 from django.contrib.auth.models import User
-from apps.recipe_api.models.category import Category
+from apps.recipes.models.category import Category
 
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
