+++ apps/recipes/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from apps.recipe_api import views
+from apps.recipes import views
 
 urlpatterns = [
     # Categories endpoints
